Introduce trait `\Icinga\Module\Notifications\Test\DbTestBackends`
